requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
werkzeug>=3.0.0
# TensorFlow/Keras for training and inference (use Python 3.11)
tensorflow>=2.20.0
# Production WSGI server for Windows
waitress>=2.1.2